Fridays homework second commit

+++ main/java/Bedroom.java
@@ -2,18 +2,19 @@
 
 public class Bedroom extends Room {
 
-    private Double roomRate;
+    private int roomRate;
     private String roomNumber;
     private String roomType;
 
-    public Bedroom(String guest, int capacity, Double roomRate, String roomNumber, String roomType) {
+
+    public Bedroom(String guest, int capacity, int roomRate, String roomNumber, String roomType){
         super(guest, capacity);
         this.roomRate = roomRate;
         this.roomNumber = roomNumber;
         this.roomType = roomType;
     }
 
-    public double getRoomRate(){
+    public int getRoomRate(){
         return roomRate;
     }
 
@@ -24,4 +25,14 @@ public class Bedroom extends Room {
     public String getRoomType(){
         return roomType;
     }
+
+//    public RoomType getType(){
+//        return this.roomType;
+//    }
+
+//    public int getTypeFromEnum() {
+//        return this.roomType;
+//    }
+
+
 }
